Rename style files
