+++ src/TODO
@@ -12,3 +12,4 @@
 [] social icons with hover state
 [] optimize icons & svgs
 [] email me for cv
+[] extend the work logos to own page
