+++ src/components/hero/hero.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import * as S from './styles';
-import { Frame, Title, SubTitle } from '../pages/styles';
+import * as S from './hero.styles';
+import { Frame, Title, SubTitle } from '../pages/homePage.styles';
 // import balloonImg from './balloon.png';
 import balloonx2Img from './balloonx2.png';
 import portraitImg from './portrait.png';
